+++ farmaceutica/include/Cliente.h
@@ -6,21 +6,21 @@
 class Cliente : public Persona
 {
 private:
-    int _cuil;
+    long long _cuil;
     int _idObraSocial;
     bool _eliminado;
 
 public:
-    Cliente(int c = 0, const char* n = "", const char* a = "",
+    Cliente(long long c = 0, const char* n = "", const char* a = "",
             int o = 0, const char* d = "", const char* m = "",
             const char* t = "", bool e = false);
 
     // Getters y setters
-    int getCuil() const;
+    long long getCuil() const;
     int getIdObraSocial() const;
     bool getEliminado() const;
 
-    void setCuil(int c);
+    void setCuil(long long c);
     void setIdObraSocial(int o);
     void setEliminado(bool e);
 };
